pyproject.toml
[tool.poetry]
name = "sparrow"
version = "0.1.0"
description = ""
authors = ["lemon <lp951195660@163.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.0.1"
flask-jwt-extended = "^4.2.3"
flask-sqlalchemy = "^2.5.1"
flask-socketio = "^5.1.0"
flask-wtf = "^0.15.1"
flask-cors = "^3.0.10"
python-dotenv = "^0.18.0"
simplejson = "^3.17.3"
mysqlclient = "^2.0.3"
gunicorn = "^20.1.0"
celery = "^5.1.2"
tablib = "^3.0.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
watchdog = "^2.1.3"
flower = "^1.0.0"
ipython = "^7.25.0"
flask-shell-ipython = "^0.4.1"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
